mp_dagster_collector-main/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==14.0
dagster==1.10.19
dagster-graphql==1.10.19
dagster-pipes==1.10.19
dagster-postgres==0.26.19
dagster-webserver==1.10.19
dagster_shared==1.10.19
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
et_xmlfile==2.0.0
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.72.1
grpcio-health-checking==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
multidict==6.4.4
narwhals==1.44.0
numpy==2.3.1
openai==1.93.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
plotly==6.2.0
propcache==0.3.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.2
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
selenium==4.34.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
structlog==25.4.0
tabulate==0.9.0
telegram==0.0.1
tomli==2.2.1
tomlkit==0.13.3
toposort==1.10
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
universal_pathlib==0.2.6
uvicorn==0.34.3
uvloop==0.21.0
watchdog==5.0.3
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
